Moved Map, and fixed Schedule bug
