+++ src/frontend/Pages/Map/components/Marker.ts
