+++ src/frontend/Pages/Map/components/Routes.ts
@@ -1,11 +1,11 @@
 import Resources from "src/backend/Resources.ts";
 import Schedule from "src/backend/Schedule.ts";
-import Vehicle from "./elements/Vehicle.ts";
+import Vehicle from "../elements/Vehicle.ts";
 import URL from "src/backend/URL.ts";
-import Route from "./elements/Route.ts";
+import Route from "../elements/Route.ts";
 import Realtime from "src/backend/Realtime.ts";
 import Peak from "src/backend/Peak.ts";
-import Stop from "./elements/Stop.ts";
+import Stop from "../elements/Stop.ts";
 
 
 namespace Routes {
