+++ src/frontend/Pages/Map/elements/Path.ts
