+++ src/frontend/Pages/Map/elements/Vehicle.ts
