+++ src/frontend/Pages/Map/elements/abstracts/InfoWindow.ts
