+++ src/frontend/Pages/Map/elements/abstracts/InfoWindowElement.ts
