+++ src/frontend/Pages/Map/elements/abstracts/Primative.ts
