+++ src/frontend/Pages/Pages.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Route, Routes } from "react-router-dom";
 
-import Map from '../Map/Map.tsx';
+import Map from './Map/Map.tsx';
 import About from './About/About.tsx';
 import Schedules from './Schedule/Schedules.tsx';
 
-function Pages() {
+export default function Pages() {
     return (
         <Routes>
             <Route path="/" element={<Map />} />
@@ -14,5 +14,3 @@ function Pages() {
         </Routes>
     )
 }
-
-export default Pages;
